Sua validate

+++ do_an/admin/delete/process_delete_admin.php
@@ -1,9 +1,8 @@
 <?php 
-	$id = $_POST['id'];
-
 	require_once '../../root/connect.php';
+	$id = mysqli_real_escape_string($ket_noi,$_POST['id']);
 
-	$delete_sql = "delete from adm_list where id = '$id'";
+	$delete_sql = "delete from adm_list where (id = '$id') and (access == 0)";
 	mysqli_query($ket_noi,$delete_sql);
 	mysqli_close($ket_noi);
 	header('location:../admin_view.php?link=admin');
